% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table-simLists.R
\name{as.data.table.simLists}
\alias{as.data.table.simLists}
\title{Coerce elements of a \code{simLists} object to a \code{data.table}}
\usage{
\method{as.data.table}{simLists}(
  x,
  keep.rownames = FALSE,
  ...,
  vals,
  objectsFromSim = NULL,
  objectsFromOutputs = NULL
)
}
\arguments{
\item{x}{An R object.}

\item{keep.rownames}{Default is \code{FALSE}. If \code{TRUE}, adds the input object's names as a separate column named \code{"rn"}. \code{keep.rownames = "id"} names the column \code{"id"} instead. For lists and when calling \code{data.table()}, names from the first named vector are extracted and used as row names, similar to \code{data.frame()} behavior.}

\item{...}{Additional arguments. Currently unused.}

\item{vals}{A (named) list of object names to extract from each
\code{simList}, or a named list of quoted expressions to calculate for each \code{simList},
or a mix of character and quoted expressions.}

\item{objectsFromSim}{Character vector of objects to extract from the simLists. If
omitted, it will extract all objects from each simList in order to calculate the
\code{vals}. This may have a computational cost. If \code{NA}, then no objects will be
accessed from the \code{simList}. Objects identified here will only be as they are in
the \code{simList}, i.e., at \code{end(sim)}.}

\item{objectsFromOutputs}{List of (named) character vectors of objects to load from the
\code{outputs(sim)} prior to evaluating \code{vals}. If there already is an object
with that same name in the \code{simList}, then it will be overwritten with
the object loaded from \code{outputs(sim)}. If there are many objects with the
same name, specifically from several \code{saveTime} values in the \code{outputs(sim)},
these will all be loaded, one at a time, \code{vals} evaluated one at a time, and
each of the values will be returned from each \code{saveTime}.
A column, \code{saveTime}, will be part of the returned \code{data.table}.
For cases where more than one object is required at a given
\code{saveTime}, all should be identified here, without time specified.
This function will take all identified objects from the same time period.}
}
\value{
This returns a \code{data.table} class object.
}
\description{
This is particularly useful to build plots using the \pkg{tidyverse}, e.g., \pkg{ggplot2}.
Ported here from the now-unmaintained \code{SpaDES.experiment} package.
}
\details{
See examples.
}
