% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{as_path}
\alias{as_path}
\title{Render a scenario (or list of them) as the canonical output path.}
\usage{
as_path(x, pre = "outputs")
}
\arguments{
\item{x}{A scenario, or anything coercible via \code{\link[=as_scenario]{as_scenario()}}.}

\item{pre}{Path prefix (default \code{"outputs"}).}
}
\description{
Render a scenario (or list of them) as the canonical output path.
}
