% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{as_tarname}
\alias{as_tarname}
\title{Render a scenario as an upload tar filename.}
\usage{
as_tarname(x, ext = ".tar.gz")
}
\arguments{
\item{x}{A scenario, or anything coercible via \code{\link[=as_scenario]{as_scenario()}}.}

\item{ext}{File extension (default \code{".tar.gz"}).}
}
\description{
Render a scenario as an upload tar filename.
}
