% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmux.R
\name{assessDoneInFigure}
\alias{assessDoneInFigure}
\title{Assess simulation status from PNG outputs}
\usage{
assessDoneInFigure(
  runName,
  timeout_min = 20,
  statusCalculate = getOption("spades.statusCalculate")
)
}
\arguments{
\item{runName}{Directory containing the figures/hists}

\item{timeout_min}{Threshold for inactivity (e.g., 20)}

\item{statusCalculate}{A quoted expression to compute job status from output files.
Defaults to \code{getOption("spades.statusCalculate", NULL)}.}
}
\description{
Assess simulation status from PNG outputs
}
