% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\title{Create or modify a project's \code{DESCRIPTION} file}
\usage{
description(fields = list(), snapshot = NULL)
}
\arguments{
\item{fields}{named list of \code{DESCRIPTION} fields and their values.}

\item{snapshot}{character. path to a \code{\link[Require:pkgSnapshot]{Require::pkgSnapshot()}} file.}
}
\value{
\code{NULL} invisibly.
Invoked for side effect of writing a DESCRIPTION file to the project directory,
and printing the resulting DESCCRIPTION file to screen.
}
\description{
A simple wrapper around \code{usethis::use_description()} to write
a \code{DESCRIPTION} file in the root project directory.
This file can be used to record project metadata, including package version
information from the current project library.
}
\note{
fields Imports and Remotes will be automatically populated based on \code{snapshot}
(or, if \code{snapshot = NULL}, the currently installed packages).
}
\examples{
tmpdir <- file.path(tempdir(), "example_project") |>
  Require::checkPath(create = TRUE)
cwd <- setwd(tmpdir)

description(fields = list(
  Title = "My Project: It Does Cool Stuff",
  Description = paste(
    "My project does so many cool things.",
    "It's so useful for all the things."
  ),
  `Authors@R` = "c(
    person('First', 'Last', , 'email@email.com', role = 'aut')
  )",
  Version = "0.0.1",
  Language = "en-CA",
  License = "MIT",
  Depends = paste0("R (== 4.2)", collapse = ",\n    ")
))

setwd(cwd)
unlink(tmpdir, recursive = TRUE)
}
