% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmux.R
\name{.assess_sim_visual_status}
\alias{.assess_sim_visual_status}
\title{Assess simulation status from PNG outputs}
\usage{
.assess_sim_visual_status(
  runName,
  timeout_min = 20,
  logPath = getOption("spades.project.tmuxLogs", "logs/tmuxStatus")
)
}
\arguments{
\item{runName}{Directory containing the figures/hists}

\item{timeout_min}{Threshold for inactivity (e.g., 20)}
}
\description{
Assess simulation status from PNG outputs
}
