% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{.libPathDefault}
\alias{.libPathDefault}
\title{SpaDES.project default .libPaths() directory}
\usage{
.libPathDefault(name)
}
\arguments{
\item{name}{A text string. When used in \code{setupProject}, this is the \code{projectName}}
}
\value{
A path where the packages will be installed.
}
\description{
For a given \code{name}, this will return the default library for packages.
}
