% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{.sync_loop_internal}
\alias{.sync_loop_internal}
\title{Sync queue to Google Sheets (Package Internal)}
\usage{
.sync_loop_internal(
  queue_path,
  ss_id,
  email,
  cache_path,
  runNameLabel,
  doneAndFinishedTimeIndicator = getOption("spades.doneAndFinishedTimeIndicator"),
  activeRunningPath = getOption("spades.activeRunningPath"),
  interval = 300
)
}
\description{
Sync queue to Google Sheets (Package Internal)
}
\keyword{internal}
