% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileEdit.R
\name{.fileEdit}
\alias{.fileEdit}
\title{Open a file for editing}
\usage{
.fileEdit(file, verbose = getOption("Require.verbose", 1L))
}
\arguments{
\item{file}{Character string giving the file path to open.}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}
}
\value{
Invoked for its side effect of opening a file for editing.
}
\description{
RStudio's \code{file.edit} behaves differently than \code{utils::file.edit}.
The workaround is to have the user manually open the file if they are using
RStudio, as suggested in the RStudio support ticket at
\url{https://support.rstudio.com/hc/en-us/community/posts/206011308-file-edit-vs-utils-file-edit}.
}
\author{
Alex Chubaty
}
\keyword{internal}
