% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{findProjectPath}
\alias{findProjectPath}
\alias{findProjectName}
\title{Find the project root directory}
\usage{
findProjectPath()

findProjectName()
}
\value{
\code{findProjectPath} returns an absolute path;
\code{findProjectName} returns the basename of the path.
}
\description{
Searches from current working directory for and Rstudio project file
or git repository, falling back on using the current working directory.
}
