% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModule.R
\name{getGithubFile}
\alias{getGithubFile}
\title{A simple way to get a Github file, authenticated}
\usage{
getGithubFile(
  gitRepoFile,
  overwrite = FALSE,
  destDir = ".",
  verbose = getOption("Require.verbose")
)
}
\arguments{
\item{gitRepoFile}{Character string that follows the convention
\emph{GitAccount/GitRepo@Branch/File}, if @Branch is omitted, then it will be
assumed to be \code{master} or \code{main}.}

\item{overwrite}{A logical vector of same length (or length 1) \code{gitRepo}.
If \code{TRUE}, then the download will delete any
existing folder with the same name as the \code{repository}
provided in \code{gitRepo}}

\item{destDir}{A directory to put the file that is to be downloaded.}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, the return object will have an
attribute: \code{attr(.., "Require")} which has lots of information about the
processes of the installs.}
}
\description{
This can be used within e.g., the \code{options} or \code{params} arguments for
\code{setupProject} to get a ready-made file for a project.
}
\examples{
filename <- getGithubFile("PredictiveEcology/LandWeb@development/01b-options.R",
                          destDir = Require::tempdir2())
}
\seealso{
\link{getModule}
}
