% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{user}
\alias{user}
\alias{machine}
\alias{node}
\title{Helpers to develop easier to understand code.}
\usage{
user(username = NULL)

machine(machinename = NULL)

node(machinename = NULL)
}
\arguments{
\item{username}{A character string of a username.}

\item{machinename}{A character string, which will be used as a partial match via
\code{grep}, so the entire machine name is not necessary. A user can use \code{regex} if
needed, e.g., \code{"^machine1"} will match \code{"machine15"} and \code{"machine12"}, but not
\code{"thisIs_machine1"}.}
}
\value{
\code{user} returns a logical indicating whether the current user matches
the supplied \code{username}.

\code{machine} returns a logical indicating whether the current machine name
\code{Sys.info()[["nodename"]]} is matched by \code{machinename}.
}
\description{
A set of lightweight helpers that are often not strictly necessary, but they
make code easier to read.
}
\details{
\code{node} is an alias for \code{machine}
}
