% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLists-class.R
\name{initialize,simLists-method}
\alias{initialize,simLists-method}
\title{Generate a \code{simLists} object}
\usage{
\S4method{initialize}{simLists}(.Object, ...)
}
\arguments{
\item{.Object}{A \code{simList} object.}

\item{...}{Optional Values passed to any or all slot}
}
\description{
Given the name or the definition of a class, plus optionally data to be
included in the object, \code{new} returns an object from that class.
}
