% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-template.R
\name{newProjectCode}
\alias{newProjectCode}
\alias{newProjectCode,character,character,character,logical-method}
\alias{newProjectCode,character,character,missing,missing-method}
\alias{newProjectCode,character,character,missing,logical-method}
\alias{newProjectCode,character,character,character,missing-method}
\title{Create new module code file}
\usage{
newProjectCode(name, path, type, open, modules = NULL, ...)

\S4method{newProjectCode}{character,character,character,logical}(name, path, type, open = interactive(), modules = NULL, ...)

\S4method{newProjectCode}{character,character,missing,missing}(name, path, type, open, modules = NULL, ...)

\S4method{newProjectCode}{character,character,missing,logical}(name, path, type, open, modules = NULL, ...)

\S4method{newProjectCode}{character,character,character,missing}(name, path, type, open, modules = NULL, ...)
}
\arguments{
\item{name}{project name (name of project directory)}

\item{path}{path to directory in which to create the project directory}

\item{type}{character string giving the project type. Possible values include:
\code{"basic"} (default), and \code{"advanced"}.}

\item{open}{Logical. Should the new project file be opened after creation?
Default \code{TRUE} in an interactive session.}

\item{modules}{A character vector of modules to download and put into the project
(see \code{SpaDES.install::getModule} for how to specify)}

\item{...}{Additional arguments. Currently only the following are implemented:
\describe{
  \item{\code{overwrite}}{
    If modules are specified, should a new module be downloaded even
    if it already exists, i.e., overwrite (and destroy).the existing.
    Default is \code{FALSE}
  }
  \item{\code{pkgPath}}{
    Path to project's package directory.
    Defaults to \file{<projectName>_packages} in the project's parent directory.
  }
}}
}
\description{
Create new module code file
}
\author{
Alex Chubaty
}
