% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{normPath}
\alias{normPath}
\alias{normPath,character-method}
\alias{normPath,list-method}
\alias{normPath,NULL-method}
\alias{normPath,missing-method}
\title{Normalize filepath}
\usage{
normPath(path)

\S4method{normPath}{character}(path)

\S4method{normPath}{list}(path)

\S4method{normPath}{`NULL`}(path)

\S4method{normPath}{missing}()
}
\arguments{
\item{path}{A character vector of filepaths.}
}
\value{
Character vector of cleaned up filepaths.
}
\description{
Checks the specified filepath for formatting consistencies:
\enumerate{
\item use slash instead of backslash;
\item do tilde etc. expansion;
\item remove trailing slash.
}
}
