% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notify_slack}
\alias{notify_slack}
\title{Send slack notification about completed simulation}
\usage{
notify_slack(runName, channel)
}
\arguments{
\item{runName}{character. The simulation run name.}

\item{channel}{character. The channel name in which to post the message.}
}
\description{
Requires the suggested \pkg{slackr} package to be installed, and configured.
(You need a webhook setup, and a config file at `~/.slackr`. See their package documentation.)
}
