% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibilityReceipt.R
\name{projectSessionInfo}
\alias{projectSessionInfo}
\title{Project session info}
\usage{
projectSessionInfo(prjDir = NULL)
}
\arguments{
\item{prjDir}{path to project directory}
}
\description{
Get detailed information about the software environment being used for a project,
including git repository information, externally installed spatial library info,
R session info, and timestamp.
}
\details{
Can be displayed inside an Rmarkdown document using \verb{[reproducibilityReceipt()]},
saved inside the project directory, or added to a SpaDES \code{simList} to improve
reproducibility of these workflows.
}
\examples{
\dontrun{
## get project session info for current project
projectSessionInfo()

## replace default session info in a simList
mySimOut <- SpaDES.core::simInitAndSpades()
mySimOut@.xData[["._sessionInfo"]] <- projectSessionInfo()
}
}
