% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibilityReceipt.R
\name{reproducibilityReceipt}
\alias{reproducibilityReceipt}
\alias{gitInfo}
\alias{submoduleInfo}
\alias{sessInfo}
\alias{spatialLibs}
\alias{timestamp}
\title{Reproducibility receipt for Rmarkdown documents}
\usage{
reproducibilityReceipt(
  prjDir = NULL,
  title = "Reproducibility receipt",
  writeTo = NULL
)

gitInfo(prjDir = NULL)

submoduleInfo(prjDir = NULL)

sessInfo(verbose = getOption("Require.verbose", 1L))

spatialLibs()

timestamp()
}
\arguments{
\item{prjDir}{path to project directory}

\item{title}{Header title for the inserted details section.}

\item{writeTo}{If provided, an markdown filename to write to (e.g., \verb{outputs/<runName>/INFO.md}).
File path is assumed to be relative to \code{prjDir}.}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, the return object will have an
attribute: \code{attr(.., "Require")} which has lots of information about the
processes of the installs.}
}
\description{
Insert git repository and R session info into Rmarkdown documents.
Based on suggestions in a Twitter thread by Miles McBain
(\url{https://twitter.com/MilesMcBain/status/1263272935197782016?s=20}).
}
\details{
Add the following to your Rmd files (without the backslashes):

\verb{
-----

\```{r details, echo=FALSE}
pemisc::reproducibilityReceipt()
\```
}
}
\seealso{
\code{\link{projectSessionInfo}}
}
