% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{setProjPkgDir}
\alias{setProjPkgDir}
\title{Set the package directory for a project}
\usage{
setProjPkgDir(lib.loc = "packages")
}
\arguments{
\item{lib.loc}{The folder for installing packages inside of}
}
\description{
This function will create a sub-folder of the \code{lib.loc} directory that
is based on the R version and the platform, as per the standard R package directory
naming convention
}
