% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{setupFiles}
\alias{setupFiles}
\title{Parse a list of (possibly remote) R / config files}
\usage{
setupFiles(
  files,
  paths,
  envir = parent.frame(),
  verbose = getOption("Require.verbose", 1L)
)
}
\arguments{
\item{files}{A vector or list of files to parse. These can be remote github.com files.}

\item{paths}{a list with named elements, specifically, \code{modulePath}, \code{projectPath},
\code{packagePath} and all others that are in \code{SpaDES.core::setPaths()}
(i.e., \code{inputPath}, \code{outputPath}, \code{scratchPath}, \code{cachePath}, \code{rasterTmpDir}).
Each of these has a sensible default, which will be overridden but any user
supplied values.
See \link{setup}.}

\item{envir}{The environment where \code{setupProject} is called from. Defaults to
\code{parent.frame()} which should be fine in most cases and user shouldn't need
to set this}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}
}
\value{
\code{setupFiles} a named list with each element that was parsed.
}
\description{
Convenience helper, intended primarily for interactive use, that parses
each file (local path or \code{github.com} URL with \verb{@branch} notation) into a
named list.
}
\details{
\code{setupFiles} is a convenience function intended for interactive use to verify the files being parsed.
This is similar to \code{parse}, but each element must be a named list or a named object, such as a function.
It uses the same specification for \url{https://github.com}
files as \code{setupProject}, i.e., using \code{@} for branch.

\if{html}{\out{<div class="sourceCode">}}\preformatted{setupFiles("PredictiveEcology/PredictiveEcology.org@main/tutos/castorExample/params.R")
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=setupProject]{setupProject()}} for the high-level wrapper, \link{setup_family} for an overview.
}
