% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{setupGitIgnore}
\alias{setupGitIgnore}
\title{Add \code{packagePath} and/or \code{modulePath} to the project's \code{.gitignore}}
\usage{
setupGitIgnore(
  paths,
  gitignore = getOption("SpaDES.project.gitignore", TRUE),
  verbose
)
}
\arguments{
\item{paths}{a list with named elements, specifically, \code{modulePath}, \code{projectPath},
\code{packagePath} and all others that are in \code{SpaDES.core::setPaths()}
(i.e., \code{inputPath}, \code{outputPath}, \code{scratchPath}, \code{cachePath}, \code{rasterTmpDir}).
Each of these has a sensible default, which will be overridden but any user
supplied values.
See \link{setup}.}

\item{gitignore}{Logical. Only has an effect if the \code{paths$projectPath}
is a git repositories without submodules. This case is ambiguous what a user
wants. If \code{TRUE}, the default, then \code{paths$modulePath} will be added to
the \code{.gitignore} file. Can be controled with \code{options(SpadES.project.gitignore = ...)}.}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}
}
\value{
\code{setupGitIgnore} is run for its side effects, i.e., adding either \code{paths$packagePath}
and/or \code{paths$modulePath} to the
\code{.gitignore} file. It will check whether \code{packagePath} is located inside the
\code{paths$projectPath} and will add this folder to the \code{.gitignore} if \code{TRUE}.
If the project is a git repository with git submodules, then it will add nothing else.
If the project is a git repository without git submodules, then the \code{paths$modulePath}
will be added to the \code{.gitignore} file. It is assumed that these modules are
used in a \verb{read only} manner.
}
\description{
Helper that keeps the \code{.gitignore} of a project under git control in sync
with the project's resolved paths.
}
\details{
\code{setupGitIgnore} will add the relevant paths to \code{.gitignore}.
}
\seealso{
\code{\link[=setupProject]{setupProject()}} for the high-level wrapper, \link{setup_family} for an overview.
}
