% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{setupStudyArea}
\alias{setupStudyArea}
\title{Resolve a study area from a \code{studyArea} spec via \code{geodata::gadm()}}
\usage{
setupStudyArea(
  studyArea,
  paths,
  envir = parent.frame(),
  callingEnv = sys.frame(-2),
  verbose = getOption("Require.verbose", 1L)
)
}
\arguments{
\item{studyArea}{Optional. If a list, it will be passed to
\code{geodata::gadm}. To specify a country other than the default \code{"CAN"},
the list must have a named element, \code{"country"}. All other named elements
will be passed to \code{gadm}. 2 additional named elements can be passed for
convenience, \code{subregion = "..."}, which will be grepped with the column
\code{NAME_1}, and \code{epsg = "..."}, so a user can pass an \code{epsg.io} code to
reproject the \code{studyArea}. Default is \code{NULL}, i.e., no study area is built.
See examples.}

\item{paths}{a list with named elements, specifically, \code{modulePath}, \code{projectPath},
\code{packagePath} and all others that are in \code{SpaDES.core::setPaths()}
(i.e., \code{inputPath}, \code{outputPath}, \code{scratchPath}, \code{cachePath}, \code{rasterTmpDir}).
Each of these has a sensible default, which will be overridden but any user
supplied values.
See \link{setup}.}

\item{envir}{The environment where \code{setupProject} is called from. Defaults to
\code{parent.frame()} which should be fine in most cases and user shouldn't need
to set this}

\item{callingEnv}{The environment from which the function was called. Defaults to \code{sys.frame(-2)}
which represents the case where the inner \verb{setup*} functions are called inside
\code{setupProject}, which was called by a user.}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}
}
\value{
\code{setupStudyArea} will return an \code{sf} class object coming from \code{geodata::gadm},
with subregion specification as described in the \code{studyArea} argument.
}
\description{
Convenience wrapper that returns an \code{sf} polygon for the requested
country / subregion using \code{geodata::gadm()}.
}
\details{
\code{setupStudyArea} calls \verb{[geodata::gadm()]} to get an \code{sf} polygon or set of polygons
of a country or a subdivision of a country. The user can pass a named \code{list} of character elements
that match entries in the columns "NAME_1" and "NAME_2" of the \code{sf} object. If passing
\code{NAME_2}, the user must pass \code{level = 2}. If passing \code{NAME_1} and \code{level = 2} all subdivision polygons under
\code{NAME_1} will be returned, which can be useful to explore subdivision names.
\code{setupStudyArea} only uses \code{inputPath} within its \code{paths} argument, which will
be passed to \code{path} argument of \code{gadm}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{setupStudyArea(list(NAME_1 = "Alberta", "NAME_2" = "Division No. 17", level = 2))
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=setupProject]{setupProject()}} for the high-level wrapper, \link{setup_family} for an overview.
}
