% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-project-package.R
\docType{package}
\name{SpaDES.project-package}
\alias{SpaDES.project}
\alias{SpaDES.project-package}
\title{Project templates using \code{SpaDES}}
\description{
Quickly setup 'SpaDES' project directories and add modules using templates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spades-project.predictiveecology.org/}
  \item \url{https://github.com/PredictiveEcology/SpaDES.project}
  \item Report bugs at \url{https://github.com/PredictiveEcology/SpaDES.project/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

}
