% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitGitRepo.R
\name{splitGitRepo}
\alias{splitGitRepo}
\title{Split a GitHub repo into its parts}
\usage{
splitGitRepo(gitRepo)
}
\arguments{
\item{gitRepo}{A character string in the form `GitAccount/GitRepo@Branch` or `@Commit`}
}
\description{
copied from \pkg{SpaDES.install}
}
\keyword{internal}
