% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.teardownProject}
\alias{.teardownProject}
\title{Helpers for cleanup of global state in examples and tests}
\usage{
.teardownProject(prjPaths, origLibPaths)
}
\arguments{
\item{prjPaths}{character vector of paths to be removed}

\item{origLibPaths}{character string giving the original library path to be restored}
}
\value{
NULL. Invoked for its side effects.
}
\description{
\enumerate{
\item remove project library directory created using \code{setupProject()};
\item remove project paths created using \code{setupProject};
\item restore original library paths.
}
}
\note{
not intended to be called by users
}
