% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{tmux_mirror_queue_to_sheets}
\alias{tmux_mirror_queue_to_sheets}
\title{Mirror local queue to Google Sheets}
\usage{
tmux_mirror_queue_to_sheets(queue_path, ss_id, sheet_name = "Status")
}
\arguments{
\item{queue_path}{Path to the local tmux_queue.rds}

\item{ss_id}{The Google Sheet ID (from the URL)}

\item{sheet_name}{The name of the tab to write to}
}
\description{
Mirror local queue to Google Sheets
}
