% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmux.R
\name{tmux_refresh_queue_status}
\alias{tmux_refresh_queue_status}
\title{Refresh and Assess Queue Status from Simulation Outputs}
\usage{
tmux_refresh_queue_status(
  queue_path,
  timeout_min = 20,
  runNameLabel,
  heartbeatFolder = NULL,
  logPath = getOption("spades.project.tmuxLogs")
)
}
\arguments{
\item{queue_path}{Character. Absolute path to the \code{experiment_queue.rds} file.}

\item{timeout_min}{Numeric. Minutes of inactivity before a task is
considered stale. Defaults to 20.}
}
\value{
A data.frame (the updated queue), invisibly.
As a side effect, updates the RDS file on disk.
}
\description{
Scans the simulation output directories (defined by \code{runNameLabel}) to assess
current status based on file timestamps and visual content of PNGs.
If a PNG has not been updated for a specified timeout, the task is marked
as "FINISHED" (if red pixels are detected) or "INTERRUPTED" (if no red
is detected).
}
\examples{
\dontrun{
# Assessment of all simulations in the current project
tmux_refresh_queue_status("experiment_queue.rds", timeout_min = 30)
}
}
