% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmux.R
\name{tmux_spawn_workers_from_df}
\alias{tmux_spawn_workers_from_df}
\title{Deprecated: use experimentTmux()}
\usage{
tmux_spawn_workers_from_df(...)
}
\value{
Invisibly returns worker pane IDs.
}
\description{
Deprecated: use experimentTmux()
}
