% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitGitRepo.R
\name{urlExists}
\alias{urlExists}
\title{check if url exists}
\usage{
urlExists(address)
}
\description{
Copied from \url{http://memosisland.blogspot.com/2012/03/check-url-existance-with-r.html}
}
\keyword{internal}
