% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmux.R
\name{activeRunningPathForTmux}
\alias{activeRunningPathForTmux}
\title{Log path for default tmux status}
\usage{
activeRunningPathForTmux(
  activeRunningPath = NULL,
  queue_path,
  prefix = "logs",
  suffix = queue_path
)
}
\value{
The default path.
}
\description{
Just a default path.
}
