% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment3}
\alias{experiment3}
\title{Run Parallel Experiments for projects that use \code{setupProject}}
\usage{
experiment3(
  expt,
  file = "global.R",
  preRunSetupProject = "paths",
  logFiles = list(expt, "time"),
  clearSimEnv = TRUE,
  saveSimToDisk = FALSE,
  tmuxName = file,
  delay = 60,
  ...
)
}
\arguments{
\item{expt}{A data.frame/data.table/list of parameters for the experiments.
Each column name must be an object that the \code{setupProject} is expecting
in the \code{.GlobalEnv}. Each row in the \code{expt} should have a single value (e.g.,
\code{character} or \code{numeric}) that will be assigned to an object named with
the column name, in the \code{.GlobalEnv}, so that the \code{setupProject} will
find it. If there is not a column named \code{.iter}, then this function will add
it, filling the column with \code{seq(NROW(expt))}. If there is not a column
named \code{.runName}, then this function will add it, filling the column with
\code{apply(expt, 1, paste(..., collapse = "_"))}, i.e. the concatenate of the
values of the \code{expt} row. If there is not a column named \code{.logFile}, then
this function will create it, filling the values with
\code{paste0(.runName, "_", format(Sys.time()))}.}

\item{file}{Character string specifying the path to the R script
(default: \code{"global.R"}). This script is expected to contain a \code{setupProject}
call. It is also expected to have very little done before the \code{setupProject}
call, including no loading of libraries (other than \code{setupProject} and its
dependencies).}

\item{preRunSetupProject}{Logical or \code{character}. Passed to \code{upTo}
in \code{preRunSetupProject}. Default is \code{packages}, so that key packages are
installed for use. This will also include \code{paths}.}

\item{clearSimEnv}{Logical. Default \code{TRUE}. This has the potential to be VERY
large. Returning it from all the threads or cores may be very inefficient. In
general, the user should define the \code{outputs} that they want so that individual
objects are saved. It will not delete any \code{dot} object, i.e., objects that start
with a \code{.}.}

\item{saveSimToDisk}{Logical or character. Default \code{FALSE}. If \code{TRUE}, then this will
save the entire \code{simList} (before any clearing via \code{clearSimEnv}, i.e, with
all objects intact) using \code{saveSimList}. The object will be saved with the name:
\code{file.path(paths$outputPath, paste0("sim_", .runName))} unless this parameter is a character
string, in which case it will be \code{file.path(paths$outputPath, paste0(saveSimToDisk, .runName))}.
In these cases, the \verb{paths$} is coming from the \code{preRunSetupProject} that returns
the paths created with \code{setupProject}. Alternatively the user can pass a
character vector with the full path of the individual simList objects to save.
If supplying a character vector of names, it must be same length as the \code{NROW(expt)},
and it must be explicit about the file extension, either \code{.rds} or \code{.qs2} for each
filename.}

\item{tmuxName}{Character string. This command will print a message with a tmux
command that can be used by a user to follow the multiple logs in a bash
command prompt. The name of the tmux window will be this.}

\item{...}{Any other objects to be made available to the \code{global.R}. These will
be added to the \code{.GlobalEnv} of the workers.}
}
\value{
Invisibly returns the result of \code{furrr::future_pmap()}.
Currently, the inner function returns \code{NULL} for each iteration.
}
\description{
Executes a set of experiments in parallel using \code{furrr::future_pmap()}.
Optionally, it can run a pre-setup routine by sourcing and evaluating
part of a script (e.g., \code{global.R}) before starting the experiments.
One important reason to do this is to parse the components of the \code{setupProject}
that define the \code{paths}, so that the \code{paths} can be used within the
experiment, e.g., to define log paths.
}
\details{
The function uses \code{furrr::future_pmap()} to run experiments in parallel.
Each experiment:
\itemize{
\item Creates a unique log file named using \code{.ELFind} and the start time.
\item Sources the \code{global.R} script in a local environment.
\item Captures messages, warnings, and errors, writing them to the log file.
}

Logging is handled via \code{withCallingHandlers()}, ensuring that all
non-empty messages, warnings, and errors are appended to the log file.
}
\examples{
\dontrun{
# Example experiment list
expt_list <- list(
  list(iter = 1, .rep = 1, .ELFind = "A"),
  list(iter = 2, .rep = 1, .ELFind = "B")
)

# Run experiments without pre-run setup
experiment3(expt = expt_list)

# Run experiments with pre-run setup
experiment3(file = "global.R", expt = expt_list)
}

}
