% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModule.R
\name{splitGitRepo}
\alias{splitGitRepo}
\alias{getGitHubFile}
\alias{isWindows}
\alias{messageDF}
\title{non-exported objects and functions from other packages}
\usage{
splitGitRepo(gitRepo)

getGitHubFile(
  pkg,
  filename = "DESCRIPTION",
  purge = getOption("Require.purge", FALSE),
  verbose = getOption("Require.verbose")
)

isWindows()

messageDF(df, round, verbose = getOption("Require.verbose"), verboseLevel = 1)
}
\description{
non-exported objects and functions from other packages
}
\keyword{internal}
