% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{messageDF}
\alias{messageDF}
\alias{messageVerbose}
\alias{downloadRepo}
\alias{.spatialPkgs}
\title{non-exported objects and functions from other packages}
\format{
An object of class \code{character} of length 9.
}
\usage{
messageDF(df, round, verbose = getOption("Require.verbose"), verboseLevel = 1)

messageVerbose(..., verbose = getOption("Require.verbose"), verboseLevel = 1)

downloadRepo(
  gitRepo,
  subFolder,
  overwrite = FALSE,
  destDir = ".",
  verbose = getOption("Require.verbose")
)

.spatialPkgs
}
\description{
non-exported objects and functions from other packages
}
\keyword{datasets}
\keyword{internal}
