% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{splitGitRepo}
\alias{splitGitRepo}
\alias{getGitHubFile}
\alias{isWindows}
\alias{messageDF}
\alias{.spatialPkgs}
\title{non-exported objects and functions from other packages}
\format{
An object of class \code{character} of length 9.
}
\usage{
splitGitRepo(gitRepo)

getGitHubFile(
  pkg,
  filename = "DESCRIPTION",
  purge = getOption("Require.purge", FALSE),
  verbose = getOption("Require.verbose")
)

isWindows()

messageDF(df, round, verbose = getOption("Require.verbose"), verboseLevel = 1)

.spatialPkgs
}
\description{
non-exported objects and functions from other packages
}
\keyword{datasets}
\keyword{internal}
