% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{messageDF}
\alias{messageDF}
\alias{yellow}
\alias{whereInStack}
\alias{blue}
\alias{messageVerbose}
\alias{compareVersion2}
\alias{downloadRepo}
\alias{splitGitRepo}
\alias{downloadFileMasterMainAuth}
\alias{isGitHub}
\alias{.spatialPkgs}
\title{non-exported objects and functions from other packages}
\format{
An object of class \code{character} of length 9.
}
\usage{
messageDF(df, round, verbose = getOption("Require.verbose"), verboseLevel = 1)

yellow(...)

whereInStack(obj)

blue(...)

messageVerbose(..., verbose = getOption("Require.verbose"), verboseLevel = 1)

compareVersion2(version, versionSpec, inequality)

downloadRepo(
  gitRepo,
  subFolder,
  overwrite = FALSE,
  destDir = ".",
  verbose = getOption("Require.verbose")
)

splitGitRepo(gitRepo, default = "PredictiveEcology", masterOrMain = NULL)

downloadFileMasterMainAuth(
  url,
  destfile,
  need = "HEAD",
  verbose = getOption("Require.verbose"),
  verboseLevel = 2
)

isGitHub(pkg, filenames)

.spatialPkgs
}
\description{
non-exported objects and functions from other packages
}
\keyword{datasets}
\keyword{internal}
