% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDESCRIPTION.R
\name{makeDESCRIPTIONproject}
\alias{makeDESCRIPTIONproject}
\alias{makeDESCRIPTION}
\title{Make DESCRIPTION file(s) from SpaDES module metadata}
\usage{
makeDESCRIPTIONproject(
  modules,
  modulePath,
  projectPath = ".",
  singleDESCRIPTION = TRUE,
  package = "Project",
  title = "Project",
  description = "Project",
  version = "1.0.0",
  authors = Sys.info()["user"],
  write = TRUE,
  verbose = getOption("Require.verbose")
)

makeDESCRIPTION(
  modules,
  modulePath,
  projectPath = ".",
  singleDESCRIPTION = FALSE,
  package,
  title,
  date,
  description,
  version,
  authors,
  write = TRUE,
  verbose,
  metadataList,
  ...
)
}
\arguments{
\item{modules}{A character vector of module names}

\item{modulePath}{Character. The path with modules, usually \code{modulePath()} or \code{paths$modulePath}}

\item{projectPath}{Character. Only used if \code{singleDESCRIPTION = TRUE}}

\item{singleDESCRIPTION}{Logical. If \code{TRUE}, there be only one DESCRIPTION file written
for all modules, i.e., all reqdPkgs will be trimmed for redundancies and put into the
single project-level DESCRIPTION file.}

\item{package}{The name inserted into the "Package" entry in DESCRIPTION}

\item{title}{The string inserted into the "Title" entry in DESCRIPTION}

\item{description}{The string inserted into the "Description" entry in DESCRIPTION}

\item{version}{The string inserted into the "Version" entry in DESCRIPTION}

\item{authors}{The string inserted into the "Authors" entry in DESCRIPTION}

\item{write}{Logical. If \code{TRUE}, then it will write the DESCRIPTION file either in
the \code{modulePath} (if \code{singleDESCRIPTION = FALSE}) or \code{projectPath}
(if \code{singleDESCRIPTION = TRUE})}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}

\item{date}{Date to enter into DESCRIPTION file. Defaults to \code{Sys.Date()}}

\item{metadataList}{The parsed source code from a module. Must include \code{defineModule} metadata.}

\item{...}{Currently not used.}
}
\description{
Make DESCRIPTION file(s) from SpaDES module metadata
}
