% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-template.R
\name{newProject}
\alias{newProject}
\alias{newProject,character,character,character,logical-method}
\alias{newProject,character,character,missing,missing-method}
\alias{newProject,character,character,character,missing-method}
\alias{newProject,character,character,missing,logical-method}
\title{Create new SpaDES project}
\usage{
newProject(name, path, type, open, ...)

\S4method{newProject}{character,character,character,logical}(name, path, type, open, ...)

\S4method{newProject}{character,character,missing,missing}(name, path, type, open, ...)

\S4method{newProject}{character,character,character,missing}(name, path, type, open, ...)

\S4method{newProject}{character,character,missing,logical}(name, path, type, open, ...)
}
\arguments{
\item{name}{project name (name of project directory)}

\item{path}{path to directory in which to create the project directory}

\item{type}{character string giving the project type. Possible values include:
\code{"basic"} (default), and \code{"advanced"}.}

\item{open}{Logical. Should the new project file be opened after creation?
Default \code{TRUE} in an interactive session.}

\item{...}{Additional arguments. Currently only the following are implemented:
\describe{
  \item{\code{pkgPath}}{
    Path to project's package directory.
    Defaults to \file{<projectName>_packages} in the project's parent directory.
  }
}}
}
\description{
Initialize a project with subdirectories \file{cache/}, \file{modules/},
\file{inputs/}, \file{outputs/}, and \code{setPaths} accordingly.
If invoked from Rstudio, will also create a new Rstudio project file.
}
\details{
This extends the basic SpaDES project template provided by \pkg{SpaDES.core}.
}
\examples{
\dontrun{
## use basic project template
myProjDir <- newProject("myProject", tempdir())

## use advanced project setup
myAdvProjDir <- newProject("myAdvProject", tempdir(), type = "advanced")
}
}
