% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyAreaPlotting.R
\name{plotSAs}
\alias{plotSAs}
\alias{plotSAsLeaflet}
\title{Plot studyArea** and rasterToMatch** with ggplot2 or leaflet}
\usage{
plotSAs(
  ll,
  ...,
  saCols = c("purple", "blue", "green", "red"),
  title,
  rasterToMatchLabel = "Stand Age",
  rasterToMatchPalette = "muted",
  country = "CAN",
  latlong = FALSE,
  minArea = 7e+11
)

plotSAsLeaflet(
  ll,
  ...,
  saCols = c("purple", "blue", "green", "red"),
  title = "Study Areas",
  rasterToMatchLabel = "Stand Age",
  rasterToMatchPalette = "muted"
)
}
\arguments{
\item{ll}{Any list-like object with named elements. Names must include at least
one that starts with \code{studyArea} or \code{rasterToMatch}. Thus any of the permutations
like \code{studyAreaLarge} or \code{rasterToMatchPSP} all are fine.}

\item{...}{Any objects to plot. Currently, they must be named arguments, and they must
have prefixes \code{studyArea} or \code{rasterToMatch} to be visualized.}

\item{saCols}{A vector of same length as number of \verb{studyArea**} objects, that defines
the studyArea polygon boundary colours. These will be used in sequence from largest
to smallest in polygon area.}

\item{title}{The main title for the ggplot2 object. Defaults to one or both of
"studyArea" and "rasterToMatch" or their plurals.}

\item{rasterToMatchLabel}{Used in rasterToMatch legend}

\item{rasterToMatchPalette}{A palette to be used for colour scheme in rasterToMatch plotting.
Can be any that work with \code{tidyterra::whitebox.colors}.}

\item{country}{The country for jurisdiction boundaries; defaults to "CAN". Passed to
\code{geodata::gadm}}

\item{latlong}{Logical. Should all layers be converted to \code{latlong} for \code{plotSAs} prior
to plotting. This means that "North will be up"; this could be slow for large rasters.
This happens by default with \code{plotSAsLeaflet} and can't be turned off.}

\item{minArea}{In m^2. This is the minimium area for the entire plot. If this is too
small then the legislative boundaries may not appear. The area covered by the plot
will the maximum of the studyArea** or rasterToMatch** and this \code{minArea} value.}
}
\value{
Run primarily for side effects. \code{plotSAs} plots (and returns) a \code{ggplot2} object.
\code{plotSAsLeaflet} creates a leaflet page in a viewer (if using Rstudio).
}
\description{
Plot all studyArea** and rasterToMatch** objects within a list-like object.
}
