% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibilityReceipt.R
\name{reproducibilityReceipt}
\alias{reproducibilityReceipt}
\alias{gitInfo}
\alias{submoduleInfo}
\alias{sessInfo}
\alias{spatialLibs}
\alias{timestamp}
\title{Reproducibility receipt for Rmarkdown documents}
\usage{
reproducibilityReceipt(title = "Reproducibility receipt", prjDir = NULL)

gitInfo(prjDir = NULL)

submoduleInfo(prjDir = NULL)

sessInfo()

spatialLibs()

timestamp()
}
\arguments{
\item{title}{Header title for the inserted details section.}
}
\description{
Insert git repository and R session info into Rmarkdown documents.
Based on suggestions in a Twitter thread by Miles McBain
(\url{https://twitter.com/MilesMcBain/status/1263272935197782016?s=20}).
}
\details{
Add the following to your Rmd files (without the backslashes):

\verb{
-----

\```{r details, echo=FALSE}
pemisc::reproducibilityReceipt()
\```
}
}
