% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibilityReceipt.R
\name{reproducibilityReceipt}
\alias{reproducibilityReceipt}
\title{Reproducibility receipt for Rmarkdown documents}
\usage{
reproducibilityReceipt(title = "Reproducibility receipt")
}
\arguments{
\item{title}{Header title for the inserted details section.}
}
\description{
Insert git repository and R session info into Rmarkdown documents.
Based on suggestions in a Twitter thread by Miles McBain
(\url{https://twitter.com/MilesMcBain/status/1263272935197782016?s=20}).
}
\details{
Add the following to your Rmd files:

\verb{
-----
<details closed>
<summary> <span title='Click to Expand'> Reproducibility receipt </span> </summary>

```r

$`Git repository`
Local:    development /home/achubaty/Documents/GitHub/PredictiveEcology/SpaDES.project
Remote:   development @ origin (ssh://git@github.com/PredictiveEcology/SpaDES.project.git)
Head:     [ed72821] 2022-09-12: tweak DESCRIPTION

$`External spatial libraries`
          GEOS           GDAL         proj.4 GDAL_with_GEOS     USE_PROJ_H 
       "3.9.1"        "3.3.2"        "7.2.1"         "true"         "true" 
          PROJ 
       "7.2.1" 

$`R session info`
- Session info ---------------------------------------------------------------
 setting  value
 version  R version 4.2.1 (2022-06-23)
 os       Ubuntu 20.04.5 LTS
 system   x86_64, linux-gnu
 ui       X11
 language en_CA:en
 collate  C
 ctype    en_CA.UTF-8
 tz       America/Edmonton
 date     2022-09-13
 pandoc   2.5 @ /usr/bin/pandoc

- Packages -------------------------------------------------------------------
 ! package        * version     date (UTC) lib source
   amc              0.2.1       2021-09-28 [1] Github (achubaty/amc@15c5229)
   base64enc        0.1-4       2021-06-18 [1] local
   bit              4.0.4       2020-08-04 [1] RSPM (R 4.2.0)
   bit64            4.0.5       2020-08-30 [1] RSPM (R 4.2.0)
   bitops           1.0-7       2021-04-24 [1] RSPM (R 4.2.0)
   blob             1.2.3       2022-04-10 [1] RSPM (R 4.2.0)
   cachem           1.0.6       2021-08-19 [1] RSPM (R 4.2.0)
   callr            3.7.2       2022-08-22 [1] RSPM
   class            7.3-20      2022-01-13 [2] CRAN (R 4.1.2)
   classInt         0.4-7       2022-06-10 [1] RSPM (R 4.2.0)
   cli              3.4.0       2022-09-08 [1] RSPM (R 4.2.0)
   clipr            0.8.0       2022-02-22 [1] RSPM (R 4.2.0)
   codetools        0.2-18      2020-11-04 [2] CRAN (R 4.0.3)
   commonmark       1.8.0       2022-03-09 [1] RSPM (R 4.2.0)
   crayon           1.5.1       2022-03-26 [1] RSPM (R 4.2.0)
   data.table       1.14.2      2021-09-27 [1] RSPM (R 4.2.0)
   DBI              1.1.3       2022-06-18 [1] RSPM (R 4.2.0)
   desc             1.4.2       2022-09-08 [1] RSPM (R 4.2.0)
   details          0.3.0       2022-03-27 [1] RSPM (R 4.2.0)
   devtools         2.4.4       2022-07-20 [1] RSPM (R 4.2.0)
   digest           0.6.29      2021-12-01 [1] RSPM (R 4.2.0)
   dismo            1.3-8       2022-07-18 [1] RSPM (R 4.2.0)
   dplyr            1.0.10      2022-09-01 [1] RSPM (R 4.2.0)
   e1071            1.7-11      2022-06-07 [1] RSPM (R 4.2.0)
   ellipsis         0.3.2       2021-04-29 [1] RSPM (R 4.2.0)
   evaluate         0.16        2022-08-09 [1] RSPM (R 4.2.0)
   fansi            1.0.3       2022-03-24 [1] RSPM (R 4.2.0)
   fasterize        1.0.3       2020-07-27 [1] RSPM (R 4.2.0)
   fastmap          1.1.0       2021-01-25 [1] RSPM (R 4.2.0)
   fpCompare        0.2.4       2022-08-15 [1] RSPM (R 4.2.0)
   fs               1.5.2       2021-12-08 [1] RSPM (R 4.2.0)
   generics         0.1.3       2022-07-05 [1] RSPM (R 4.2.0)
   git2r            0.30.1      2022-03-16 [1] RSPM (R 4.2.0)
   glue             1.6.2       2022-02-24 [1] RSPM (R 4.2.0)
   here             1.0.1       2020-12-13 [1] RSPM (R 4.2.0)
   htmltools        0.5.3       2022-07-18 [1] RSPM (R 4.2.0)
   htmlwidgets      1.5.4       2021-09-08 [1] RSPM (R 4.2.0)
   httpuv           1.6.6       2022-09-08 [1] RSPM (R 4.2.0)
   httr             1.4.4       2022-08-17 [1] RSPM (R 4.2.0)
   KernSmooth       2.23-20     2021-05-03 [2] CRAN (R 4.0.5)
   knitr            1.40        2022-08-24 [1] RSPM
   later            1.3.0       2021-08-18 [1] RSPM (R 4.2.0)
   lattice          0.20-45     2021-09-22 [2] CRAN (R 4.2.0)
   lifecycle        1.0.2       2022-09-09 [1] RSPM (R 4.2.0)
   lobstr           1.1.2       2022-06-22 [1] RSPM (R 4.2.0)
   magrittr         2.0.3       2022-03-30 [1] RSPM (R 4.2.0)
   memoise          2.0.1       2021-11-26 [1] RSPM (R 4.2.0)
   mime             0.12        2021-09-28 [1] RSPM (R 4.2.0)
   miniUI           0.1.1.1     2018-05-18 [1] RSPM (R 4.2.0)
   pemisc           0.0.4       2022-09-13 [1] https://predictiveecology.r-universe.dev (R 4.2.1)
   peutils          0.1.0       2021-06-01 [1] Github (PredictiveEcology/peutils@d40d97c)
   pillar           1.8.1       2022-08-19 [1] RSPM
   pkgbuild         1.3.1       2021-12-20 [1] RSPM (R 4.2.0)
   pkgconfig        2.0.3       2019-09-22 [1] RSPM (R 4.2.0)
   pkgload          1.3.0       2022-06-27 [1] RSPM (R 4.2.0)
   png              0.1-8       2021-06-18 [1] local
   prettyunits      1.1.1       2020-01-24 [1] RSPM (R 4.2.0)
   processx         3.7.0       2022-07-07 [1] RSPM (R 4.2.0)
   profvis          0.3.7       2020-11-02 [1] RSPM (R 4.2.0)
   promises         1.2.0.1     2021-02-11 [1] RSPM (R 4.2.0)
   proxy            0.4-27      2022-06-09 [1] RSPM (R 4.2.0)
   ps               1.7.1       2022-06-18 [1] RSPM (R 4.2.0)
   purrr            0.3.4       2020-04-17 [1] RSPM (R 4.2.0)
   R6               2.5.1       2021-08-19 [1] RSPM (R 4.2.0)
   raster           3.5-29      2022-08-14 [1] CRAN (R 4.2.1)
   Rcpp             1.0.9       2022-07-08 [1] RSPM (R 4.2.0)
   RCurl            1.98-1.8    2022-07-30 [1] RSPM (R 4.2.0)
   remotes          2.4.2       2021-11-30 [1] RSPM (R 4.2.0)
   reproducible     1.2.10.9001 2022-09-13 [1] https://predictiveecology.r-universe.dev (R 4.2.1)
   Require          0.1.1.9009  2022-09-09 [1] local
   rgdal            1.5-32      2022-05-09 [1] CRAN (R 4.2.0)
   rlang            1.0.5       2022-08-31 [1] RSPM (R 4.2.0)
   roxygen2         7.2.1       2022-07-18 [1] RSPM (R 4.2.0)
   rprojroot        2.0.3       2022-04-02 [1] RSPM (R 4.2.0)
   RSQLite          2.2.17      2022-09-10 [1] RSPM (R 4.2.0)
   rstudioapi       0.14        2022-08-22 [1] RSPM
   sessioninfo      1.2.2       2021-12-06 [1] RSPM (R 4.2.0)
   sf               1.0-8       2022-07-14 [1] CRAN (R 4.2.1)
   shiny            1.7.2       2022-07-19 [1] RSPM (R 4.2.0)
   sp               1.5-0       2022-06-05 [1] RSPM
 P SpaDES.project * 0.0.6       2022-06-21 [?] Github (PredictiveEcology/SpaDES.project@7a9dd06)
   stringi          1.7.8       2022-07-11 [1] RSPM (R 4.2.0)
   stringr          1.4.1       2022-08-20 [1] RSPM
   terra            1.6-7       2022-08-07 [1] CRAN (R 4.2.1)
   tibble           3.1.8       2022-07-22 [1] RSPM (R 4.2.0)
   tidyselect       1.1.2       2022-02-21 [1] RSPM (R 4.2.0)
   units            0.8-0       2022-02-05 [1] RSPM (R 4.2.0)
   urlchecker       1.0.1       2021-11-30 [1] RSPM (R 4.2.0)
   usethis          2.1.6       2022-05-25 [1] RSPM (R 4.2.0)
   utf8             1.2.2       2021-07-24 [1] RSPM (R 4.2.0)
   vctrs            0.4.1       2022-04-13 [1] RSPM (R 4.2.0)
   whisker          0.4         2019-08-28 [1] RSPM (R 4.2.0)
   withr            2.5.0       2022-03-03 [1] RSPM (R 4.2.0)
   xfun             0.33        2022-09-12 [1] RSPM (R 4.2.0)
   xml2             1.3.3       2021-11-30 [1] RSPM (R 4.2.0)
   xtable           1.8-4       2019-04-21 [1] RSPM (R 4.2.0)

 [1] /home/achubaty/R/x86_64-pc-linux-gnu-library/4.2
 [2] /usr/lib/R/library
 [3] /usr/local/lib/R/site-library
 [4] /usr/lib/R/site-library

 P -- Loaded and on-disk path mismatch.

------------------------------------------------------------------------------

$Timestamp
[1] "2022-09-13 10:24:31 MDT"

```

</details>
<br>
}
}
