% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmux.R
\name{runNextWorker}
\alias{runNextWorker}
\title{Run one queued job (claim-next semantics) in the current R session.}
\usage{
runNextWorker(
  queue_path,
  global_path,
  on_interrupt = c("requeue", "fail"),
  heartbeat_interval_s = 60,
  runNameLabel = quote(colnames(q)[1:2]),
  folderWithDoneIndicator = getOption("spades.folderWithDoneIndicator"),
  folderWithIterInFilename = getOption("spades.folderWithIterInFilename"),
  activeRunningPath = getOption("spades.activeRunningPath")
)
}
\arguments{
\item{queue_path}{character; path to the queue \code{.rds}}

\item{global_path}{character; script to source for the job}

\item{on_interrupt}{"requeue" or "fail". If the sourced script is interrupted, either requeue or mark as FAILED.}

\item{heartbeat_interval_s}{numeric; seconds between heartbeats while the job runs}

\item{runNameLabel}{A quoted expression (possibly of \code{q}, which is the result of \code{q <- readRDS(queue_path)}).
Default is the first 2 column names of \code{q}. These will be concatenated and used as
labels for various things including the \code{activeRunningPath} file(s).}
}
\value{
"ok" | "interrupt" | "empty" (if no pending work found); used by runWorkerLoop()
}
\description{
Run one queued job (claim-next semantics) in the current R session.
}
