% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths2.R
\name{setProjPkgDir}
\alias{setProjPkgDir}
\title{Set the package directory for a project}
\usage{
setProjPkgDir(lib.loc = "packages", verbose = getOption("Require.verbose", 1L))
}
\arguments{
\item{lib.loc}{The folder for installing packages inside of}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}
}
\description{
This function will create a sub-folder of the \code{lib.loc} directory that
is based on the R version and the platform, as per the standard R package directory
naming convention
}
