% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{setupModules}
\alias{setupModules}
\title{Download (or git clone) SpaDES modules into the project's \code{modulePath}}
\usage{
setupModules(
  name,
  paths,
  modules,
  inProject,
  useGit = getOption("SpaDES.project.useGit", FALSE),
  overwrite = FALSE,
  envir = parent.frame(),
  callingEnv = sys.frame(-2),
  gitUserName,
  verbose = getOption("Require.verbose", 1L),
  dots,
  defaultDots,
  updateRprofile = getOption("SpaDES.project.updateRprofile", TRUE),
  ...
)
}
\arguments{
\item{name}{Optional. If supplied, the name of the project. If not supplied, an
attempt will be made to extract the name from the \code{paths[["projectPath"]]}.
If this is a GitHub project, then it should indicate the full Github
repository and branch name, e.g., \code{"PredictiveEcology/WBI_forecasts@ChubatyPubNum12"}}

\item{paths}{a list with named elements, specifically, \code{modulePath}, \code{projectPath},
\code{packagePath} and all others that are in \code{SpaDES.core::setPaths()}
(i.e., \code{inputPath}, \code{outputPath}, \code{scratchPath}, \code{cachePath}, \code{rasterTmpDir}).
Each of these has a sensible default, which will be overridden but any user
supplied values.
See \link{setup}.}

\item{modules}{a character vector of modules to pass to \code{getModule}. These
should be one of: simple name (e.g., \code{fireSense}) which will be searched for locally
in the \code{paths[["modulePath"]]}; or a GitHub repo with branch (\code{GitHubAccount/Repo@branch} e.g.,
\code{"PredictiveEcology/Biomass_core@development"}); or a character vector that identifies
one or more module folders (local or GitHub) (not the module .R script).
If the entire project is a git repository,
then it will not try to re-get these modules; instead it will rely on the user
managing their git status outside of this function.
For convenience, these can also be 2 other url formats:
\enumerate{
\item the raw.githubusercontent.com url that points to the main module file or the folder e.g.,
\code{"https://raw.githubusercontent.com/PredictiveEcology/Biomass_core/refs/heads/main/Biomass_core.R"}
\item The github.com url used for cloning a git repository, with optional "@branch" specified:
\code{"https://github.com/PredictiveEcology/Biomass_speciesParameters.git@development"}
See \link{setup}.
}}

\item{inProject}{A logical. If \code{TRUE}, then the current directory is
inside the \code{paths[["projectPath"]]}.}

\item{useGit}{(if not FALSE, then experimental still). There are two levels at which a project
can use GitHub, either the \code{projectPath} and/or the \code{modules}. Any given
project can have one or the other, or both of these under git control. If "both",
then this function will assume that git submodules will be used for the \code{modules}.
A logical or \code{"sub"} for \emph{submodule}. If \code{"sub"}, then this function
will attempt to clone the identified \code{modules} \emph{as git submodules}. This will only
work if the \code{projectPath} is a git repository. If the project is already a git
repository because the user has set that up externally to this function call, then
this function will add the \code{modules} as git submodules. If it is not already,
it will use \verb{git clone} for each module. After git clone or submodule add are run,
it will run \verb{git checkout} for the named branch and then \verb{git pull}
to get and change branch for each module, according to its specification in
\code{modules}. If \code{FALSE}, this function  will download modules with \code{getModules}.
NOTE: \emph{CREATING} A
GIT REPOSITORY AT THE PROJECT LEVEL AND SETTING MODULES AS GIT SUBMODULES IS
EXPERIMENTAL. IT IS FINE IF THE PROJECT HAS BEEN MANUALLY SET UP TO BE
A GIT REPOSITORY WITH SUBMODULES: THIS FUNCTION WILL ONLY EVALUTE PATHS. This can
be set with the \code{option(SpaDES.project.useGit = xxx)}. Default is
\code{getOption("SpaDES.project.useGit", FALSE)}, i.e., \code{FALSE} unless that option is set.}

\item{overwrite}{Logical vector or character vector, however, only \code{getModule} will respond
to a vector of values. If length-one \code{TRUE}, then all files that were previously downloaded
will be overwritten throughout the sequence of \code{setupProject} -- including those downloaded via \code{sideEffects}.
If a length > 1 logical or character vector, these will be passed to \code{getModule}: only the named
modules will be overwritten or the logical vector of the modules.
NOTE: if length > 1, no other file specified anywhere in \code{setupProject} will be
overwritten except a module matching the vector \code{names()} (because
only \code{setupModules} is currently responsive to a vector). To have fine grained control,
a user can just manually delete a file, then rerun. Default is
\code{getOption("SpaDES.project.overwrite", FALSE)}, i.e., \code{FALSE} unless that option is set.}

\item{envir}{The environment where \code{setupProject} is called from. Defaults to
\code{parent.frame()} which should be fine in most cases and user shouldn't need
to set this}

\item{callingEnv}{The environment from which the function was called. Defaults to \code{sys.frame(-2)}
which represents the case where the inner \verb{setup*} functions are called inside
\code{setupProject}, which was called by a user.}

\item{gitUserName}{The GitHub account name. Used with git clone git@github.com:\emph{gitHuserName}/name}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}

\item{dots}{Any other named objects passed as a list a user might want for other elements.}

\item{defaultDots}{A named list of any arbitrary R objects.
These can be supplied to give default values to objects that
are otherwise passed in with the \code{...}, i.e., not specifically named for these
\verb{setup*} functions. If named objects are supplied as top-level arguments, then
the \code{defaultDots} will be overridden. This can be particularly useful if the
arguments passed to \code{...} do not always exist, but rely on external e.g., batch
processing to optionally fill them. See examples.}

\item{updateRprofile}{Logical. Should the \code{paths$packagePath} be set in the \code{.Rprofile}
file for this project. Note: if \code{paths$packagePath} is within the \code{tempdir()},
then there will be a warning, indicating this won't persist. If the user is
using \code{Rstudio} and the \code{paths$projectPath} is not the root of the current
Rstudio project, then a warning will be given, indicating the .Rprofile may not
be read upon restart. Default is \code{getOption("SpaDES.project.updateRprofile", TRUE)},
i.e., \code{TRUE} unless that option is set.}

\item{...}{further named arguments that acts like \code{objects}, but a different
way to specify them. These can be anything. The general use case
is to create the \code{objects} that are would be passed to
\code{SpaDES.core::simInit}, or \code{SpaDES.core::simInitAndSpades},
(e.g. \code{studyAreaName} or \code{objects}) or additional objects to be passed to the simulation
(in older versions of \code{SpaDES.core}, these were passed as a named list
to the \code{objects} argument). \strong{Order matters}. These are sequentially evaluated,
and also any arguments that are specified before the named arguments
e.g., \code{name}, \code{paths}, will be evaluated prior to any of the named arguments,
i.e., "at the start" of the \code{setupProject}.
If placed after the first named argument, then they will be evaluated at the
end of the \code{setupProject}, so can access all the packages, objects, etc.}
}
\value{
\code{setupModules} is run for its side effects, i.e., downloads modules and puts them
into the \code{paths[["modulePath"]]}. It will return a named list, where the names are the
full module names and the list elemen.ts are the R packages that the module
depends on (\code{reqsPkgs})
}
\description{
Materialise the modules requested in \code{\link[=setupProject]{setupProject()}} beneath
\code{paths[["modulePath"]]}, optionally as git submodules.
}
\details{
\code{setupModules} will download all modules do not yet exist locally. The current
test for "exists locally" is simply whether the directory exists. If a user
wants to update the module, \code{overwrite = TRUE} must be set, or else the user can
remove the folder manually.
}
\seealso{
\code{\link[=setupProject]{setupProject()}} for the high-level wrapper, \link{setup_family} for an overview.
}
