% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{setupOptions}
\alias{setupOptions}
\title{Apply (and stage) project options}
\usage{
setupOptions(
  name,
  options,
  paths,
  times,
  overwrite = FALSE,
  envir = parent.frame(),
  callingEnv = sys.frame(-2),
  verbose = getOption("Require.verbose", 1L),
  dots,
  defaultDots,
  useGit = getOption("SpaDES.project.useGit", FALSE),
  updateRprofile = getOption("SpaDES.project.updateRprofile", TRUE),
  ...
)
}
\arguments{
\item{name}{Optional. If supplied, the name of the project. If not supplied, an
attempt will be made to extract the name from the \code{paths[["projectPath"]]}.
If this is a GitHub project, then it should indicate the full Github
repository and branch name, e.g., \code{"PredictiveEcology/WBI_forecasts@ChubatyPubNum12"}}

\item{options}{Optional. Either a named list to be passed to \code{options}
or a character vector indicating one or more file(s) to source,
in the order provided. These will be parsed locally (not
the \code{.GlobalEnv}), so they will not create globally accessible objects. NOTE:
\code{options} is run 2x within \code{setupProject}, once before \code{setupPaths} and once
after \code{setupPackages}. This occurs because many packages use options for their
behaviour (need them set before e.g., \code{Require::require} is run; but many packages
also change \code{options} at startup. See details.
See \link{setup}.}

\item{paths}{a list with named elements, specifically, \code{modulePath}, \code{projectPath},
\code{packagePath} and all others that are in \code{SpaDES.core::setPaths()}
(i.e., \code{inputPath}, \code{outputPath}, \code{scratchPath}, \code{cachePath}, \code{rasterTmpDir}).
Each of these has a sensible default, which will be overridden but any user
supplied values.
See \link{setup}.}

\item{times}{Optional. This will be returned if supplied; if supplied, the values
can be used in e.g., \code{params}, e.g., \code{params = list(mod = list(startTime = times$start))}.
Default (when not supplied) is \code{list(start = 0, end = 1)}.
See help for \code{SpaDES.core::simInit}.}

\item{overwrite}{Logical vector or character vector, however, only \code{getModule} will respond
to a vector of values. If length-one \code{TRUE}, then all files that were previously downloaded
will be overwritten throughout the sequence of \code{setupProject} -- including those downloaded via \code{sideEffects}.
If a length > 1 logical or character vector, these will be passed to \code{getModule}: only the named
modules will be overwritten or the logical vector of the modules.
NOTE: if length > 1, no other file specified anywhere in \code{setupProject} will be
overwritten except a module matching the vector \code{names()} (because
only \code{setupModules} is currently responsive to a vector). To have fine grained control,
a user can just manually delete a file, then rerun. Default is
\code{getOption("SpaDES.project.overwrite", FALSE)}, i.e., \code{FALSE} unless that option is set.}

\item{envir}{The environment where \code{setupProject} is called from. Defaults to
\code{parent.frame()} which should be fine in most cases and user shouldn't need
to set this}

\item{callingEnv}{The environment from which the function was called. Defaults to \code{sys.frame(-2)}
which represents the case where the inner \verb{setup*} functions are called inside
\code{setupProject}, which was called by a user.}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}

\item{dots}{Any other named objects passed as a list a user might want for other elements.}

\item{defaultDots}{A named list of any arbitrary R objects.
These can be supplied to give default values to objects that
are otherwise passed in with the \code{...}, i.e., not specifically named for these
\verb{setup*} functions. If named objects are supplied as top-level arguments, then
the \code{defaultDots} will be overridden. This can be particularly useful if the
arguments passed to \code{...} do not always exist, but rely on external e.g., batch
processing to optionally fill them. See examples.}

\item{useGit}{(if not FALSE, then experimental still). There are two levels at which a project
can use GitHub, either the \code{projectPath} and/or the \code{modules}. Any given
project can have one or the other, or both of these under git control. If "both",
then this function will assume that git submodules will be used for the \code{modules}.
A logical or \code{"sub"} for \emph{submodule}. If \code{"sub"}, then this function
will attempt to clone the identified \code{modules} \emph{as git submodules}. This will only
work if the \code{projectPath} is a git repository. If the project is already a git
repository because the user has set that up externally to this function call, then
this function will add the \code{modules} as git submodules. If it is not already,
it will use \verb{git clone} for each module. After git clone or submodule add are run,
it will run \verb{git checkout} for the named branch and then \verb{git pull}
to get and change branch for each module, according to its specification in
\code{modules}. If \code{FALSE}, this function  will download modules with \code{getModules}.
NOTE: \emph{CREATING} A
GIT REPOSITORY AT THE PROJECT LEVEL AND SETTING MODULES AS GIT SUBMODULES IS
EXPERIMENTAL. IT IS FINE IF THE PROJECT HAS BEEN MANUALLY SET UP TO BE
A GIT REPOSITORY WITH SUBMODULES: THIS FUNCTION WILL ONLY EVALUTE PATHS. This can
be set with the \code{option(SpaDES.project.useGit = xxx)}. Default is
\code{getOption("SpaDES.project.useGit", FALSE)}, i.e., \code{FALSE} unless that option is set.}

\item{updateRprofile}{Logical. Should the \code{paths$packagePath} be set in the \code{.Rprofile}
file for this project. Note: if \code{paths$packagePath} is within the \code{tempdir()},
then there will be a warning, indicating this won't persist. If the user is
using \code{Rstudio} and the \code{paths$projectPath} is not the root of the current
Rstudio project, then a warning will be given, indicating the .Rprofile may not
be read upon restart. Default is \code{getOption("SpaDES.project.updateRprofile", TRUE)},
i.e., \code{TRUE} unless that option is set.}

\item{...}{further named arguments that acts like \code{objects}, but a different
way to specify them. These can be anything. The general use case
is to create the \code{objects} that are would be passed to
\code{SpaDES.core::simInit}, or \code{SpaDES.core::simInitAndSpades},
(e.g. \code{studyAreaName} or \code{objects}) or additional objects to be passed to the simulation
(in older versions of \code{SpaDES.core}, these were passed as a named list
to the \code{objects} argument). \strong{Order matters}. These are sequentially evaluated,
and also any arguments that are specified before the named arguments
e.g., \code{name}, \code{paths}, will be evaluated prior to any of the named arguments,
i.e., "at the start" of the \code{setupProject}.
If placed after the first named argument, then they will be evaluated at the
end of the \code{setupProject}, so can access all the packages, objects, etc.}
}
\value{
\code{setupOptions} is run for its side effects, namely, changes to the \code{options()}. The
list of modified options will be added as an attribute (\code{attr(out, "projectOptions")}),
e.g., so they can be "unset" by user later.
}
\description{
Set the \code{options()} supplied to \code{\link[=setupProject]{setupProject()}} and record the prior values
so they can be restored.
}
\details{
\code{setupOptions} can handle sequentially specified values, meaning a user can
first create a list of default options, then a list of user-desired options that
may or may not replace individual values. Thus final values will be based on the
order that they are provided.
}
\seealso{
\code{\link[=setupProject]{setupProject()}} for the high-level wrapper, \link{setup_family} for an overview.
}
