% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{setupProject}
\alias{setupProject}
\title{Sets up a new or existing SpaDES project}
\usage{
setupProject(name, paths, modules, packages)
}
\arguments{
\item{name}{The name of the project; if this is a GitHub project, then it should
indicate the full Github repository and branch name, e.g.,
\code{"PredictiveEcology/WBI_forecasts@ChubatyPubNum12"}}

\item{paths}{a list with named elements, specifically, \code{modulePath}, \code{projectPath},
\code{packagePath} and all others that are in \code{SpaDES.core::setPaths()} `(
inputPath, outputPath, scratchPath, cachePath, rasterTmpDir)``}

\item{modules}{a character string of modules to pass to \code{getModule}. These
should be in the form \code{GitHubAccount/Repo@branch} e.g.,
\code{"PredictiveEcology/Biomass_core@development"}. If the project is a git repository,
then it will not try to re-get these modules; instead it will rely on the user
managing their git status outside of this function.}

\item{packages}{A vector of packages that are needed. This will be passed to
\code{Require}}
}
\description{
The main things this function does are: creates folder structures, downloads or
confirms existence of modules, install missing packages from both the modules
\code{reqdPkgs} fields and the user passed \code{packages}. See Details.
}
\details{
This function bundles a sequence of functions: \code{.libPaths()}, \code{getModule}, \code{setwd},
\code{Require}
}
\examples{
\dontrun{
setupProject("test", list(modulePath = "m"),
             modules = "PredictiveEcology/Biomass_core",
             packages = "fpCompare")
}
}
