% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupProject.R
\name{setupSideEffects}
\alias{setupSideEffects}
\title{Run side-effect scripts (e.g., authentication, custom package options)}
\usage{
setupSideEffects(
  name,
  sideEffects,
  paths,
  times,
  overwrite = FALSE,
  envir = parent.frame(),
  callingEnv = sys.frame(-2),
  verbose = getOption("Require.verbose", 1L),
  dots,
  defaultDots,
  ...
)
}
\arguments{
\item{name}{Optional. If supplied, the name of the project. If not supplied, an
attempt will be made to extract the name from the \code{paths[["projectPath"]]}.
If this is a GitHub project, then it should indicate the full Github
repository and branch name, e.g., \code{"PredictiveEcology/WBI_forecasts@ChubatyPubNum12"}}

\item{sideEffects}{Optional. This can be an expression or one or more file names or
a code chunk surrounded by \code{{...}}.
If a non-text file name is specified (e.g., \emph{not .txt or .R} currently),
these files will simply be downloaded, using their relative path as specified
in the github notation. They will be downloaded or accessed locally at that
relative path.
If these file names represent scripts (*.txt or .R), this/these will be parsed and evaluated,
but nothing is returned (i.e., any assigned objects are not returned). This is intended
to be used for operations like cloud authentication or configuration functions
that are run for their side effects only.}

\item{paths}{a list with named elements, specifically, \code{modulePath}, \code{projectPath},
\code{packagePath} and all others that are in \code{SpaDES.core::setPaths()}
(i.e., \code{inputPath}, \code{outputPath}, \code{scratchPath}, \code{cachePath}, \code{rasterTmpDir}).
Each of these has a sensible default, which will be overridden but any user
supplied values.
See \link{setup}.}

\item{times}{Optional. This will be returned if supplied; if supplied, the values
can be used in e.g., \code{params}, e.g., \code{params = list(mod = list(startTime = times$start))}.
Default (when not supplied) is \code{list(start = 0, end = 1)}.
See help for \code{SpaDES.core::simInit}.}

\item{overwrite}{Logical vector or character vector, however, only \code{getModule} will respond
to a vector of values. If length-one \code{TRUE}, then all files that were previously downloaded
will be overwritten throughout the sequence of \code{setupProject} -- including those downloaded via \code{sideEffects}.
If a length > 1 logical or character vector, these will be passed to \code{getModule}: only the named
modules will be overwritten or the logical vector of the modules.
NOTE: if length > 1, no other file specified anywhere in \code{setupProject} will be
overwritten except a module matching the vector \code{names()} (because
only \code{setupModules} is currently responsive to a vector). To have fine grained control,
a user can just manually delete a file, then rerun. Default is
\code{getOption("SpaDES.project.overwrite", FALSE)}, i.e., \code{FALSE} unless that option is set.}

\item{envir}{The environment where \code{setupProject} is called from. Defaults to
\code{parent.frame()} which should be fine in most cases and user shouldn't need
to set this}

\item{callingEnv}{The environment from which the function was called. Defaults to \code{sys.frame(-2)}
which represents the case where the inner \verb{setup*} functions are called inside
\code{setupProject}, which was called by a user.}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}

\item{dots}{Any other named objects passed as a list a user might want for other elements.}

\item{defaultDots}{A named list of any arbitrary R objects.
These can be supplied to give default values to objects that
are otherwise passed in with the \code{...}, i.e., not specifically named for these
\verb{setup*} functions. If named objects are supplied as top-level arguments, then
the \code{defaultDots} will be overridden. This can be particularly useful if the
arguments passed to \code{...} do not always exist, but rely on external e.g., batch
processing to optionally fill them. See examples.}

\item{...}{further named arguments that acts like \code{objects}, but a different
way to specify them. These can be anything. The general use case
is to create the \code{objects} that are would be passed to
\code{SpaDES.core::simInit}, or \code{SpaDES.core::simInitAndSpades},
(e.g. \code{studyAreaName} or \code{objects}) or additional objects to be passed to the simulation
(in older versions of \code{SpaDES.core}, these were passed as a named list
to the \code{objects} argument). \strong{Order matters}. These are sequentially evaluated,
and also any arguments that are specified before the named arguments
e.g., \code{name}, \code{paths}, will be evaluated prior to any of the named arguments,
i.e., "at the start" of the \code{setupProject}.
If placed after the first named argument, then they will be evaluated at the
end of the \code{setupProject}, so can access all the packages, objects, etc.}
}
\value{
\code{setupSideEffects} is run for its side effects (e.g., web authentication, custom package
options that cannot use \code{base::options}), with deliberately nothing returned to user.
This, like other parts of this function, attempts to prevent unwanted outcomes
that occur when a user uses e.g., \code{source} without being very careful about
what and where the objects are sourced to.
}
\description{
Source the side-effect scripts or expressions supplied to \code{\link[=setupProject]{setupProject()}};
nothing is returned to the user.
}
\details{
Most arguments in the family of \verb{setup*} functions are run \emph{sequentially}, even within
the argument. Since most arguments take lists, the user can set values at a first
value of a list, then use it in calculation of the 2nd value and so on. See
examples. This "sequential" evaluation occurs in the \code{...}, \code{setupSideEffects}, \code{setupOptions},
\code{setupParams} (this does not work for \code{setupPaths}) can handle sequentially
specified values, meaning a user can
first create a list of default options, then a list of user-desired options that
may or may not replace individual values. This can create hierarchies, \emph{based on
order}.
}
\seealso{
\code{\link[=setupProject]{setupProject()}} for the high-level wrapper, \link{setup_family} for an overview.
}
