% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaDES.projectOptions.R
\name{spadesProjectOptions}
\alias{spadesProjectOptions}
\title{\code{SpaDES.project} options}
\usage{
spadesProjectOptions()
}
\value{
named list of the \emph{default} package options.
}
\description{
These provide top-level, powerful settings for a comprehensive SpaDES workflow.
To see defaults, run \code{spadesProjectOptions()}.
See Details below.
}
\details{
Below are options that can be set with \code{options("spadesProject.xxx" = newValue)},
where \code{xxx} is one of the values below, and \code{newValue} is a new value to
give the option. Sometimes these options can be placed in the user's \code{.Rprofile}
file so they persist between sessions.

The following options are used, using the prefix: \code{spades}
\tabular{lcl}{
\emph{OPTION} \tab \emph{DEFAULT VALUE} \tab \emph{DESCRIPTION} \cr
\code{spades.logPath}
\tab Defaults to a subdirectory (\verb{logs/}) of the simulation output directory.
\tab The default local directory in which to look for simulation inputs.  \cr

\code{spades.inputPath}
\tab Default is a temporary directory (typically \verb{/tmp/RtmpXXX/SpaDES/inputs})
\tab The default local directory in which to look for simulation inputs.  \cr

\code{spades.modulePath} \tab \code{file.path(tempdir(), "SpaDES", "modules")})
\tab The default local directory where modules and data will be downloaded and stored.
Default is a temporary directory  \cr

\code{spades.outputPath}
\tab \code{file.path(tempdir(), "SpaDES", "outputs")}
\tab The default local directory in which to save simulation outputs.\cr

\code{spades.useRequire} \tab \code{!tolower(Sys.getenv("SPADES_USE_REQUIRE")) \%in\% "false"}
\tab : The default for that environment variable is unset, so this returns
\code{TRUE}. If this is \code{TRUE}, then during the \code{simInit} call, when packages are
identified as being required, these will be installed if missing, only if
\code{spades.useRequire} option is \code{TRUE}, otherwise, \code{simInit} will fail because
packages are not available.\cr

}
}
