% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaDES.projectOptions.R
\name{spadesProjectOptions}
\alias{spadesProjectOptions}
\title{\code{SpaDES.project} options}
\usage{
spadesProjectOptions()
}
\value{
named list of the \emph{default} options currently available.
}
\description{
These demonstrate default values for some options that can be set in
SpaDES.project.
To see defaults, run \code{spadesProjectOptions()}.
See Details below.
}
\details{
Below are options that can be set with \code{options("spades.xxx" = newValue)},
where \code{xxx} is one of the values below, and \code{newValue} is a new value to
give the option. Sometimes these options can be placed in the user's \code{.Rprofile}
file so they persist between sessions.

The following options are used, using the prefix: \code{spades}
\tabular{lcl}{
\emph{OPTION} \tab \emph{DEFAULT VALUE} \tab \emph{DESCRIPTION} \cr
\code{reproducible.cachePath}
\tab NOTE: uses \code{reproducible}. Defaults is within projectPath, with subfolder "cache"  \cr

\code{spades.inputPath}
\tab Default is within projectPath, with subfolder "inputs"  \cr

\code{spades.modulePath}
\tab Default is within projectPath, with subfolder "modules"  \cr

\code{spades.outputPath}
\tab Default is within projectPath, with subfolder "outputs"  \cr

\code{spades.packagePath}
\tab Default to \verb{.libPathDefault(<projectPath>)}  \cr

\code{spades.projectPath}
\tab Default "."  \cr

\code{spades.scratchPath}
\tab Default is within \code{tempdir()}, with subfolder \if{html}{\out{<projectPath>}}  \cr

\code{SpaDES.project.Restart}
\tab Default is FALSE. Passed to \code{Restart} argument in \code{setupProject}  \cr

\code{SpaDES.project.useGit}
\tab Default is FALSE. Passed to \code{useGit} argument in \code{setupProject}  \cr

}
}
