% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitGitRepo.R
\name{splitGitRepo}
\alias{splitGitRepo}
\title{This is copied from SpaDES.installl}
\usage{
splitGitRepo(gitRepo)
}
\arguments{
\item{gitRepo}{A character string in the form GitAccount/GitRepo@Branch or @Commit}
}
\description{
This is copied from SpaDES.installl
}
